requirements.txt
# app
# pygemstones>=0.0.13
# black>=22.0.0
pytz 
python-dateutil
pygemstones
black
pydantic
pydantic[email]
pydantic-settings
fastapi_mail
fastapi
uvicorn
requests
sqlalchemy
mysql-connector-python
throttled
apscheduler
itsdangerous
dotenv
passlib
asyncpg
greenlet
bcrypt 
cuid
openai
asyncio
python-jose[cryptography]
python-multipart
# aioredis 
redis
# fastapi-cache2
# test
pytest
pytest-asyncio
pytest-mock
pytest-cov
httpx
cloudinary
